Fix enum, new constructor for ProjectsException
