+++ src/Utils/Jalasoft.TeamUp.Projects.ProjectsException/ProjectsErrors.cs
@@ -1,0 +1,9 @@
+﻿namespace Jalasoft.TeamUp.Projects.ProjectsException
+{
+    public enum ProjectsErrors
+    {
+        InternalServerError = 500,
+        BadRequest = 400,
+        NotFound = 404,
+    }
+}
