+++ src/api/Jalasoft.TeamUp.Projects.API/Startup.cs
@@ -20,9 +20,9 @@ namespace Jalasoft.TeamUp.Projects.API
             builder.Services.AddScoped<IHealthRepository, HealthRepository>();
             builder.Services.AddScoped<IProjectsService, ProjectsService>();
 #if DEBUG
-            builder.Services.AddScoped<IRepository<Project>, ProjectsMongoDbRepository>();
-#else
             builder.Services.AddScoped<IRepository<Project>, ProjectsInMemoryRepository>();
+#else
+            builder.Services.AddScoped<IRepository<Project>, ProjectsMongoDbRepository>();
 #endif
         }
     }
